requirements.txt
Flask==2.3.3
tensorflow==2.13.0
pillow==9.2.0
numpy==1.22.0
Flask-Cors==3.0.10
gunicorn==20.1.0
